requirements.txt
djangorestframework-simplejwt==5.2.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
daphne==4.0.0
Django==4.2.4
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
gunicorn==21.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
msgpack==1.0.5
openapi-codec==1.3.2
packaging==23.1
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-decouple==3.8
python-dotenv==1.0.0
python-environ==0.4.54
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.0
service-identity==23.1.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.23.2
zope.interface==6.0
